Update index.html
Hopefully added persistence to not lose what's been scanned if the page refreshes.  Current logic has a list of 6 digit skus with a comma and the quantity needed. 

In order to know how many case packs to send I need to take the truck manifest and have AI give each sku a carton quantity (this can simply be added to for trucks going forward) and the number of case packs (for the current truck). Then when scanning, compare the number needed to the cases, if the total is less than 1 case send that case and backstock the rest, if the total is greater than a case send however many cases are needed.

How do we want the files to look? For now a basic comma and line seperated list works. Maybe truck manifest - sku,numberincasepack,numberofcasepacks, aka 069420,50,1,

That would be loaded under the "manifest" button (maybe we can automate the extraction process?) and be used as a refrence for the backstock list. Going forward the backstock list should probably change to a floorstock list